[7-DEV-PS] effacer la carte seulement lorqu'on confirme

+++ client/src/app/list-of-games-view/list-of-games-view.component.ts
@@ -34,7 +34,8 @@ export class ListOfGamesViewComponent {
     swal({
       text: "Vouler vous vraiment supprimer ce jeu ?",
       buttons: ["Annuler", "Supprimer le jeu!"],
-    }).catch((err: Error) => console.error(err));
-    this.listOfGamesService.delete(gameCard);
+    })
+    .then((confirm: boolean) => {if (confirm) {this.listOfGamesService.delete(gameCard); }})
+    .catch((err: Error) => console.error(err));
   }
 }
